BACKEND_LATEST/IICB/requirement.txt
argcomplete==1.10.0
argh==0.26.2
astroid==2.3.0
benchmark-django-rest-framework==1.0.0
certifi==2019.9.11
chardet==3.0.4
configparser==4.0.2
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.2
django-cors-headers==3.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.1.1
djangorestframework==3.9.4
gffpandas==1.2.0
gffutils==0.9
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.3
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mysql-connector-python==8.0.16
mysql-connector-python-dd==2.0.2
numpy==1.17.2
openapi-codec==1.3.2
pandas==0.24.2
protobuf==3.8.0
pybase64==1.0.0
pyfaidx==0.5.5.2
pylint==2.4.1
PyMySQL==0.9.3
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.25.6
wrapt==1.11.2
